requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
elasticsearch==7.10.0
elasticsearch-dsl==7.3.0
entrypoints==0.3
filelock==3.0.12
flake8==3.8.4
future==0.18.2
idna==2.10
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kiwisolver==1.3.1
lxml==4.6.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
meilisearch==0.13.0
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nltk==3.5
notebook==6.1.5
numpy==1.19.4
oauthlib==3.1.0
openai==0.2.6
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pipenv==2020.11.4
pkg-resources==0.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-rake==1.5.0
pytz==2020.4
pyzmq==20.0.0
qtconsole==4.7.7
QtPy==1.9.0
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
sacremoses==0.0.43
scipy==1.5.4
seaborn==0.11.0
Send2Trash==1.5.0
sentencepiece==0.1.91
six==1.15.0
soupsieve==2.0.1
terminado==0.9.1
testpath==0.4.4
tokenizers==0.9.3
toml==0.10.2
torch==1.7.0
torchvision==0.8.1
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
transformers==3.5.1
tweepy==3.9.0
typing-extensions==3.7.4.3
uritools==3.0.0
urlextract==1.1.0
urllib3==1.25.11
virtualenv==20.1.0
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
